+++ mukurobot/__init__.py
@@ -7,8 +7,6 @@ import sys
 
 __version__ = '0.2.0-dev'
 
-CURRENCY_SYMBOL = '<:Macoto:1088440100158963733>'
-
 dotenv.load_dotenv()
 
 from .client import Mukuro